# src/main/resources/static/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Swagger Rimberry API - OpenAPI 3.0
  contact:
    email: ismailrzouki7@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: https://localhost:8080/api/v1
...
security:
   - bearerAuth: []
paths:
...
  /supplier/add:
    post:
      security:
        - bearerAuth: []
      tags:
        - Supplier Management
      operationId: addSupplier
      summary: add supplier operation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierDto'
        required: true
      responses:
        '200':
          description: Supplier Successfully Added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
        '500':
          description: Failed to Add New Supplier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
  /supplier/edit:
    put:
      security:
        - bearerAuth: []
      tags:
        - Supplier Management
      operationId: editSupplier
      summary: Edit Supplier
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierDto'
        required: true
      responses:
        '200':
          description: Supplier Successfully Added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
        '500':
          description: Failed to Edit supplier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
  /supplier/delete/{supplierId}:
    delete:
      security:
        - bearerAuth: []
      tags:
        - Supplier Management
      operationId: deleteSupplier
      summary: Delete specific supplier
      parameters:
        - in: path
          name: supplierId
          schema:
            type: integer
          required: true
          description: Numeric ID of the supplier to delete
      responses:
        '200':
          description: Supplier Successfully Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
        '500':
          description: Failed to Delete product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
  /supplier/search:
    get:
      security:
        - bearerAuth: []
      tags: 
        - Supplier Management
      operationId: searchSupplier
      summary: search supplier
      parameters:
        - in: query
          name: text
          description: lookup text
          required: true
          schema:
            type: string
            example: text
      responses:
        '200':
          description: Supplier List search result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierDtoList'
            application/xml:
              schema:
                $ref: '#/components/schemas/SupplierDtoList'
        '500':
          description: Failed to Find any Supplier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SupplierDto:
      type: object
      properties:
        id:
          type: integer
          example: 344
        name:
          type: string
          example: ahmed
        address:
          type: string
          example: N 55 LOT SMARA LARACHE
        cin:
          type: string
          example: LA14481
        rib:
          type: string
          example: 1234567891234567
        source:
          type: string
          example: string
        created:
          type: string
          format: date-time
          example: null
        updated:
          type: string
          format: date-time
          example: null
        isDeleted:
          type: boolean
          example: false
    SupplierDtoList:
      type: array
      items:
        $ref: '#/components/schemas/SupplierDto'    
...
    Response:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: Exception Details from the API
        details:
          type: string
          example: "Exception in thread 'main' java.lang.NullPointerException: Oops!"